223- comments policy

+++ modules/Cyaxaress/Comment/src/Http/Controllers/CommentController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Cyaxaress\Comment\Http\Controllers;
 
 use App\Http\Controllers\Controller;
@@ -6,23 +7,36 @@ use Cyaxaress\Comment\Http\Requests\CommentRequest;
 use Cyaxaress\Comment\Models\Comment;
 use Cyaxaress\Comment\Repositories\CommentRepo;
 use Cyaxaress\Common\Responses\AjaxResponses;
+use Cyaxaress\Course\Models\Course;
+use Cyaxaress\RolePermissions\Models\Permission;
 
-class CommentController extends Controller{
+class CommentController extends Controller
+{
 
     public function index(CommentRepo $repo)
     {
+        $this->authorize('index', Comment::class);
         $comments = $repo
             ->searchBody(request("body"))
             ->searchEmail(request("email"))
             ->searchName(request("name"))
-            ->searchStatus(request("status"))
-            ->paginateParents();
+            ->searchStatus(request("status"));
+
+        if (!auth()->user()->hasAnyPermission(Permission::PERMISSION_MANAGE_COMMENTS, Permission::PERMISSION_SUPER_ADMIN)) {
+            $comments->query->whereHasMorph("commentable", [Course::class] , function ($query) {
+                return $query->where("teacher_id", auth()->id());
+            });
+        }
+
+        $comments = $comments->paginateParents();
+
         return view("Comments::index", compact("comments"));
     }
 
     public function show($comment)
     {
         $comment = Comment::query()->where("id", $comment)->with("commentable", "user", "comments")->firstOrFail();
+        $this->authorize('view', $comment);
         return view("Comments::show", compact("comment"));
     }
 
@@ -35,7 +49,7 @@ class CommentController extends Controller{
 
     public function accept($id, CommentRepo $commentRepo)
     {
-//        $this->authorize('change_confirmation_status', Course::class);
+        $this->authorize('manage', Comment::class);
         if ($commentRepo->updateStatus($id, Comment::STATUS_APPROVED)) {
             return AjaxResponses::SuccessResponse();
         }
@@ -45,7 +59,7 @@ class CommentController extends Controller{
 
     public function reject($id, CommentRepo $commentRepo)
     {
-//        $this->authorize('change_confirmation_status', Course::class);
+        $this->authorize('manage', Comment::class);
         if ($commentRepo->updateStatus($id, Comment::STATUS_REJECTED)) {
             return AjaxResponses::SuccessResponse();
         }
@@ -55,6 +69,7 @@ class CommentController extends Controller{
 
     public function destroy($id, CommentRepo $repo)
     {
+        $this->authorize('manage', Comment::class);
         $comment = $repo->findOrFail($id);
         $comment->delete();
         return AjaxResponses::SuccessResponse();
